+++ CHANGELOG.md
@@ -55,6 +55,7 @@
 * fix named let with nil [#46](https://github.com/jcubic/lips/issues/46)
 * use parser code in string->number
 * fix repr of native object instances (e.g.: `Uint8Array`)
+* fix display/write without newline in both REPLs
 
 ## 0.20.1
 ### Bug fixes